k8s: Separate CiliumClusterwideNetworkPolicy CRD
This commit separates out CiliumClusterwideNetworkPolicy (CCNP) from
v2/types.go to prepare for use with the controller-gen tool for
automatic CRD generation.
