+++ pkg/k8s/apis/cilium.io/v2/types.go
@@ -27,37 +27,6 @@ import (
 )
 
 // +genclient
-// +genclient:nonNamespaced
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-
-// CiliumClusterwideNetworkPolicy is a Kubernetes third-party resource with an modified version
-// of CiliumNetworkPolicy which is cluster scoped rather than namespace scoped.
-// +deepequal-gen=false
-type CiliumClusterwideNetworkPolicy struct {
-	*CiliumNetworkPolicy
-
-	// Status is the status of the Cilium policy rule
-	// +optional
-	// The reason this field exists in this structure is due a bug in the k8s code-generator
-	// that doesn't create a `UpdateStatus` method because the field does not exist in
-	// the structure.
-	Status CiliumNetworkPolicyStatus `json:"status"`
-}
-
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-
-// CiliumClusterwideNetworkPolicyList is a list of CiliumClusterwideNetworkPolicy objects
-// +k8s:openapi-gen=false
-// +deepequal-gen=false
-type CiliumClusterwideNetworkPolicyList struct {
-	metav1.TypeMeta `json:",inline"`
-	metav1.ListMeta `json:"metadata"`
-
-	// Items is a list of CiliumClusterwideNetworkPolicy
-	Items []CiliumClusterwideNetworkPolicy `json:"items"`
-}
-
-// +genclient
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
 
 // CiliumEndpoint is the status of a Cilium policy rule
